+++ src/test/scala/org/zouzias/spark/lucenerdd/LucenePrimitiveTypesSpec.scala
@@ -18,7 +18,6 @@ package org.zouzias.spark.lucenerdd
 
 import com.holdenkarau.spark.testing.SharedSparkContext
 import org.scalatest.{BeforeAndAfterEach, FlatSpec, Matchers}
-import org.zouzias.spark.lucenerdd.implicits.LuceneRDDImplicits._
 
 class LucenePrimitiveTypesSpec extends FlatSpec with Matchers
   with BeforeAndAfterEach
