+++ src/test/scala/org/zouzias/spark/lucenerdd/LuceneTuplesRDDSpec.scala
@@ -18,7 +18,6 @@ package org.zouzias.spark.lucenerdd
 
 import com.holdenkarau.spark.testing.SharedSparkContext
 import org.scalatest.{ FlatSpec, Matchers }
-import org.zouzias.spark.lucenerdd.implicits.LuceneRDDImplicits._
 
 class LuceneTuplesRDDSpec extends FlatSpec with Matchers with SharedSparkContext {
